Move expanded code out, hook up \x special key
