+++ pgcli/main.py
@@ -18,8 +18,9 @@ from pygments.lexers.sql import SqlLexer
 from psycopg2 import Error
 
 from .packages.tabulate import tabulate
+from .packages.expanded import expanded_table
 from .packages.pgspecial import (CASE_SENSITIVE_COMMANDS,
-        NON_CASE_SENSITIVE_COMMANDS)
+        NON_CASE_SENSITIVE_COMMANDS, is_expanded_output)
 from .pgcompleter import PGCompleter
 from .pgtoolbar import PGToolbar
 from .pgstyle import PGStyle
@@ -29,47 +30,6 @@ from .config import write_default_config, load_config
 from .key_bindings import pgcli_bindings
 
 _logger = logging.getLogger(__name__)
-
-def pad(field, total, char=u" "):
-    return field + (char * (total - len(field)))
-
-def get_separator(num, header_len, data_len):
-    total_len = header_len + data_len + 1
-
-    sep = u"-[ RECORD {0} ]".format(unicode(num))
-    if len(sep) < header_len:
-        sep = pad(sep, header_len - 1, u"-") + u"+"
-
-    if len(sep) < total_len:
-        sep = pad(sep, total_len, u"-")
-
-    return sep + u"\n"
-
-def expanded_table(rows, headers):
-    header_len = max([len(x) for x in headers])
-    max_row_len = 0
-    results = []
-
-    padded_headers = [pad(x, header_len) + u" |" for x in headers]
-    header_len += 2
-
-    for row in rows:
-        row_len = max([len(str(x)) for x in row])
-        row_result = ""
-        if row_len > max_row_len:
-            max_row_len = row_len
-
-        for item in zip(padded_headers, row):
-            row_result += item[0] + u" " + unicode(item[1]) + u"\n"
-
-        results.append(row_result)
-
-    output = ""
-    for i, result in enumerate(results):
-        output += get_separator(i, header_len, max_row_len)
-        output += result
-
-    return output
 
 @click.command()
 @click.option('-h', '--host', default='', help='Host address of the '
@@ -150,8 +110,10 @@ def cli(database, user, password, host, port):
                     _logger.debug("headers: %r", headers)
                     _logger.debug("rows: %r", rows)
                     if rows:
-                        # output.append(tabulate(rows, headers, tablefmt='fancy_grid'))
-                        output.append(expanded_table(rows, headers))
+                        if is_expanded_output():
+                            output.append(expanded_table(rows, headers))
+                        else:
+                            output.append(tabulate(rows, headers, tablefmt='psql'))
                     if status:  # Only print the status if it's not None.
                         output.append(status)
                     _logger.debug("status: %r", status)
