+++ pgcli/packages/pgspecial.py
@@ -19,6 +19,9 @@ class MockLogging(object):
         print ()
 
 #log = MockLogging()
+use_expanded_output = False
+def is_expanded_output():
+    return use_expanded_output
 
 def parse_special_command(sql):
     command, _, arg = sql.partition(' ')
@@ -687,7 +690,11 @@ def change_db(cur, arg, verbose):
     raise NotImplementedError
 
 def expanded_output(cur, arg, verbose):
-    import ipdb; ipdb.set_trace()
+    global use_expanded_output
+    use_expanded_output = not use_expanded_output
+    message = u"Expanded display is "
+    message += u"on" if use_expanded_output else u"off"
+    return [(None, None, message + u".")]
 
 CASE_SENSITIVE_COMMANDS = {
             '\?': (show_help, ['\?', 'Help on pgcli commands.']),
